requirements.txt
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.2
PyGithub>=2.1.1
python-gitlab>=4.0.0
GitPython>=3.1.40
faiss-cpu>=1.7.4
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
